refactor: `TeamsChannelItem` to ts
