--- apps/meteor/client/views/teams/contextualBar/channels/Row.js
@@ -1,13 +1,0 @@
-import React, { memo } from 'react';
-
-import TeamsChannelItem from './TeamsChannelItem';
-
-function Row({ room, onClickView, reload }) {
-	if (!room) {
-		return <TeamsChannelItem.Skeleton />;
-	}
-
-	return <TeamsChannelItem room={room} onClickView={() => onClickView(room)} reload={reload} />;
-}
-
-export default memo(Row);
