+++ apps/meteor/client/views/teams/contextualBar/channels/index.ts
@@ -1,1 +1,1 @@
-export { default } from './TeamsChannels';
+export { default } from './TeamsChannelsWithData';
